feat: decoder fixes and re-structure
