+++ lib/mqtt/index.js
@@ -13,6 +13,10 @@ function transformData(payload, serial) {
   const events = [];
   const ts = new Date().toISOString();
 
+  if (!Array.isArray(payload)) {
+    return events;
+  }
+
   for (const item of payload) {
 
     if (item.type && item.unit && item.value && item.channel) {
